Added initial function pointer type, to be one of the replacement primitive types for the built-in function type.
Currently, due to lack of variadic templates, there are multiple
primitive types for a function pointer and this will need to be
fixed in future. There are also ABI issues with the 'call'
method of the function pointer type that need to be addressed.

SEM::FunctionType has been added as part of this change, which is
an abstract representation of a function signature (this seems to
be needed in various places of the compiler). It's a type that
wraps a pointer to 'SEM::FunctionTypeData' for convenience;
SEM::Type will soon be upgraded to use a similar mechanism.