ref T39: Fixed parser to reject comparison operations in function arguments.
Syntax such as 'f(A<B, C>(1))' is currently ambiguous because it
can either be constructing a type 'A' with template arguments 'B'
and 'C', or it can be performing two separate comparisons and
passing their results to the function as two arguments.

This change fixes the parser to reject any comparisons in function
arguments that aren't bracketed, which eliminates the ambiguity.