Removed invalid assert hit due to new diagnostics.
This assert was previously valid, but the new diagnostics system
means we continue after errors have occurred, hence this assert
will be hit because it shadows a compiler error. Hence the fix is
to remove the assert, since we want to be able to continue running
Semantic Analysis after compiler errors.