Split hintResultValue out of IREmitter.
Currently 'hintResultValue' (a pointer into which the result can
be stored to avoid moves; this can be specified as NULL) is passed
to IREmitter and used by IREmitter::emitReturnAlloca().

However this logic can be complex, since the 'hintResultValue'
effectively disappears into the IREmitter constructor and
re-appears later. Furthermore, a new IREmitter has to be created
when performing a recursive call without the same (or any)
hintResultValue.

Hence this change moves 'hintResultValue' out of IREmitter so it
is now passed around as needed. This makes the code simpler and
more flexible (at the cost of an extra parameter in a few cases).