locic: Output necessary information for tests.
The CheckError tests need to see the unformatted form of old-style
error messages. They also need to see the new diagnostics even if
an old-style error is encountered. Hence this change fixes locic
to print these if it's given the '--verify' flag.

This change will become unnecessary once all old-style error
messages are upgraded to the new diagnostics system.