Allow users to specify standard library modules to use.
This adds the syntax <module-name>:<module-version> to the
compiler's command line syntax, so that any number of standard
library modules can be specified and the relevant API files and
binary files are added to the build. For example:

    locic my_source.loci std.string:0.1.0 -o my_module.bc

In this example, 'std.string' depends on 'std.memory', so the
latter will also be added to the build.