Don't throw SkipException if ImplicitCast fails.
SkipException is used to prevent Semantic Analysis continuing if
particular errors are detected; it only exists to facilitate the
previous conversion to the new diagnostics system and should
gradually be removed from the compiler.

This change removes SkipException from ImplicitCast, instead
returning a 'dummy' value with the desired destination type.