Fix method set satisfies logic to substitute selfconst.
Whenever we perform 'A : B' to check that type 'A' satisfies the
requirements specified by type 'B', we now also perform selfconst
substitution for both types.

For example, with 'const A : B' the value of selfconst is true
for 'A' and false for 'B'. If 'A' had a method returning
selfconst(T)& and B had a similar method, 'const A : B' would be
FALSE because after substitution of selfconst it would be
attempting to cast 'const(T)&' to 'T&'.