Only use method sets for interface types.
This change means that 'Type<A> : Type<B>' where Type is NOT an
interface type will just try to unify the types rather than
looking at their method sets. This is necessary because the method
sets of classes and template variables may be incomplete; adding
in the hidden methods may make the satisfies check false.

For some types (e.g. a datatype) the complete method set is known,
so it could be possible at some point to enable method set checking
for them.